More code cleanup. Moving toward a more onion like approach to the generated code instead of flat
